Move toll chat components to a new folder
